Rename minecode_pipeline to minecode_pipelines
