+++ .github/workflows/pypi-release-minecode-pipeline.yml
@@ -23,13 +23,13 @@ jobs:
         run: python -m pip install flot --user
 
       - name: Build binary wheel and source tarball
-        run:  python -m flot --pyproject pyproject-minecode_pipeline.toml --sdist --wheel --output-dir dist/
+        run:  python -m flot --pyproject pyproject-minecode_pipelines.toml --sdist --wheel --output-dir dist/
 
       - name: Publish to PyPI
         if: startsWith(github.ref, 'refs/tags')
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
-          password: ${{ secrets.PYPI_API_TOKEN_MINECODE_PIPELINE }}
+          password: ${{ secrets.PYPI_API_TOKEN_MINECODE_PIPELINES }}
 
       - name: Upload built archives
         uses: actions/upload-artifact@v4
