+++ Makefile
@@ -140,7 +140,7 @@ test_matchcode:
 test: test_purldb test_matchcode test_toolkit test_clearcode
 
 test_minecode:
-	${ACTIVATE} ${PYTHON_EXE} -m pytest -vvs minecode_pipeline
+	${ACTIVATE} ${PYTHON_EXE} -m pytest -vvs minecode_pipelines
 
 shell:
 	${MANAGE} shell
